# gen/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email:   
    name:     
    url: https://josue-lubaki.ca
  description: Springular REST API
  license:
    name: Apache License Version 2.0
  title: Springular REST API
  version: 1.0.0
servers:
- description: Base Url
  url: http://localhost:8080
...
paths:
  /auth/login:
    post:
      description: Login to the application
      operationId: authenticateUserUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginDTO'
          description: OK
        "400":
          description: Bad Request
      summary: Login
      tags:
      - Registration of accounts
      x-contentType: application/json
      x-accepts: application/json
  /auth/register:
    post:
      description: Register to the application
      operationId: registerUserUsingPOST
      parameters:
      - explode: true
        in: query
        name: apartment
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: city
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: dob
        required: false
        schema:
          format: date
          type: string
        style: form
      - explode: true
        in: query
        name: email
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: fname
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: imageURL
        required: false
        schema:
          format: binary
          type: string
        style: form
      - explode: true
        in: query
        name: lname
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: password
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: phone
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: role
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: state
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: street
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: zip
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: Successfully created an account
        "400":
          description: Bad Request
      summary: Register
      tags:
      - Registration of accounts
      x-accepts: application/json
  /auth/reset-password:
    post:
      operationId: resetPasswordUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDTO'
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Successfully reset password
      summary: Reset password
      tags:
      - Registration of accounts
      x-contentType: application/json
      x-accepts: application/json
...
  /api/v1/users/{id}:
    delete:
      operationId: deleteMyAccountUsingDELETE
      parameters:
      - description: User ID to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "204":
          description: Successfully Deleted a user
      security:
      - JWT:
        - global
      summary: Delete a user by ID
      tags:
      - User Controller
      x-accepts: application/json
    get:
      operationId: getUserUsingGET
      parameters:
      - description: User ID to consult
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
          description: Successfully retrieved a user
      security:
      - JWT:
        - global
      summary: Get a user by ID
      tags:
      - User Controller
      x-accepts: '*/*'
    put:
      operationId: updateUserUsingPUT
      parameters:
      - description: User ID to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
          description: Successfully updated a user
      security:
      - JWT:
        - global
      summary: Update a user by ID
      tags:
      - User Controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /management/api/v1/users/{id}:
    delete:
      operationId: deleteUserUsingDELETE
      parameters:
      - description: user ID to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "204":
          description: Successfully Deleted a user
      security:
      - JWT:
        - global
      summary: Delete a user by ID
      tags:
      - User Management Controller
      x-accepts: application/json
    get:
      operationId: getUserUsingGET_1
      parameters:
      - description: user ID to consult
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
          description: Successfully retrieved a user
      security:
      - JWT:
        - global
      summary: Get a user by ID
      tags:
      - User Management Controller
      x-accepts: '*/*'
    put:
      operationId: updateUserUsingPUT_1
      parameters:
      - description: user ID to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTOReq'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
          description: Successfully updated a user
      security:
      - JWT:
        - global
      summary: Update a user by ID
      tags:
      - User Management Controller
      x-contentType: application/json
      x-accepts: '*/*'
components:
  schemas:
    AddressDTO:
      example:
        zip: zip
        city: city
        street: street
        id: 0
        state: state
        apartement: apartement
      properties:
        apartement:
          type: string
        city:
          type: string
        id:
          format: int64
          type: integer
        state:
          type: string
        street:
          type: string
        zip:
          type: string
      title: AddressDTO
      type: object
...
    AuthDTO:
      example:
        password: password
        username: username
      properties:
        username:
          type: string
        password:
          type: string
      title: AuthDTO
      type: object
...
    LoginDTO:
      example:
        role: role
        id: 0
        email: email
        token: token
      properties:
        email:
          type: string
        id:
          format: int64
          type: integer
        role:
          type: string
        token:
          type: string
      title: LoginDTO
      type: object
...
    UserDTO:
      example:
        fname: fname
        address:
          zip: zip
          city: city
          street: street
          id: 0
          state: state
          apartement: apartement
        role: role
        created: 2000-01-23
        fullName: fullName
        lname: lname
        deleted: true
        phone: phone
        dob: 2000-01-23
        imageURL: imageURL
        id: 6
        updated: 2000-01-23
        email: email
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          format: date
          type: string
        deleted:
          type: boolean
        dob:
          format: date
          type: string
        email:
          type: string
        fname:
          type: string
        fullName:
          type: string
        id:
          format: int64
          type: integer
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          format: date
          type: string
      title: UserDTO
      type: object
    UserDTOReq:
      example:
        fname: fname
        lname: lname
        address:
          zip: zip
          city: city
          street: street
          id: 0
          state: state
          apartement: apartement
        deleted: true
        role: role
        phone: phone
        created: 2000-01-23
        dob: 2000-01-23
        imageURL: imageURL
        id: 0
        updated: 2000-01-23
        email: email
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          format: date
          type: string
        deleted:
          type: boolean
        dob:
          format: date
          type: string
        email:
          type: string
        fname:
          type: string
        id:
          format: int64
          type: integer
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          format: date
          type: string
      title: UserDTOReq
      type: object
    UserDTORes:
      example:
        fname: fname
        address:
          zip: zip
          city: city
          street: street
          id: 0
          state: state
          apartement: apartement
        role: role
        created: 2000-01-23
        fullName: fullName
        lname: lname
        deleted: true
        phone: phone
        dob: 2000-01-23
        imageURL: imageURL
        id: 0
        updated: 2000-01-23
        email: email
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          format: date
          type: string
        deleted:
          type: boolean
        dob:
          format: date
          type: string
        email:
          type: string
        fname:
          type: string
        fullName:
          type: string
        id:
          format: int64
          type: integer
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          format: date
          type: string
      title: UserDTORes
      type: object
...
  securitySchemes:
    JWT:
      in: header
      name: JWT
      type: apiKey