# src/main/java/ca/ghostteam/springulart/dto/LoginDTO.java
package ca.ghostteam.springulart.dto;

import lombok.Data;

/**
 * @author     
 * @version 1.0
 * @since 2022-03-25
 */
@Data
public class LoginDTO {
    private Long id;
    private String email;
    private String token;
    private String role;
    private String status;
}